[FIX] Sales channel should not update pricelist_id in invoice.
pricelist_id in Standard Odoo is not part of account.invoice. It is only added by account_invoice_pricelist
module. It might be a good idea to have additional modules that fill pricelist in invoice from sales order
when both sale and account_invoice_pricelist are installed.

account_invoice_pricelist makes sure that when prices are computed on the invoice, the pricelist on the partner
is used (actually this needs acceptance of a Pull Request, withouth this pricelist_id on invoice is only
informational).

Took the opportunity to modernize the headers and do away with excessive logging.